Overhaul of data structure and handling
With this new approach changing the duration of parts of the cycle is now possible.
example: User wants the morningslope to change from 60 minutes to 50. This is now simply done by changing `morningSlopeDuration` in the settings.
Also `minPerTimeCode` can now be changed withour recompiling the dataset. The data tables now support float values here too, as does the TimeCode class.